requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.10.4
attrs==25.3.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.2.1
celery==5.3.6
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.0
cryptography==44.0.2
deep-translator==1.11.4
deepl==1.21.1
dnspython==2.7.0
EbookLib==0.18
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
frozenlist==1.6.0
googletrans==2.4.0
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.26.0
idna==3.10
iniconfig==2.1.0
kombu==5.5.3
lxml==5.3.2
motor==3.7.0
multidict==6.4.3
packaging==25.0
passlib==1.7.4
playwright==1.51.0
pluggy==1.5.0
prompt_toolkit==3.0.51
propcache==0.3.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyee==12.1.1
pymongo==4.12.0
pytest==8.2.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
redis==5.0.1
requests==2.32.3
rsa==4.9.1
scalar_fastapi==1.0.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.0